SE置換実装の準備
- コンテキスト確定時点でEmoji、Tagを置き換え
- Emoji, Tagはあとから置き換える時齟が出ないようWYSIWYG
